Get Makefile working
This leverages julia#59108 to generate a log file with the information
needed to create a language binding. As that PR is not yet merged, this
may be subject to change.

This appears to get things working up through trimming, at which point
there are many verifier errors that need to be resolved.